log_analyzer/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-text-splitters
langchain-groq
python-multipart
